modify move and rename
